backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gunicorn==25.0.3
PyJWT==2.8.0
python-dotenv==1.2.1
redis==5.0.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.10
Werkzeug==3.1.5
jsonschema==4.23.0
